add total model count to radial charts
