+++ src/pages/overview.tsx
@@ -31,39 +31,9 @@ import {
 import PieChart from "../components/Charts/PieChart";
 import PolarBarChart from "../components/Charts/PolarBarChart";
 import RadialChart from "../components/Charts/RadialChart";
-import SunBurstChart from "../components/Charts/SunBurstChart";
-import { ageCategories } from "../utils/collapseEthnicity";
-import { capitalizeFirstLetter } from "../utils/dataUtils";
-
-const transformSunBurstData = (data: Record<string, number>) => {
-	const labels: string[] = [];
-	const values: number[] = [];
-	const parents: string[] = [];
-	let parentTotalCount = 0,
-		tempChildValues: number[] = [];
-
-	Object.entries(ageCategories).forEach(([category, categoryValues]) => {
-		// assuming all of this categories are included in our models (if not, why is it even in the dictionary?)
-		labels.push(capitalizeFirstLetter(category.toLowerCase()));
-		parents.push("");
-		parentTotalCount = 0;
-		tempChildValues = [];
-
-		Object.entries(data).forEach(([age, count]) => {
-			if (categoryValues.includes(age)) {
-				labels.push(age);
-				tempChildValues.push(count);
-				parents.push(capitalizeFirstLetter(category.toLowerCase()));
-				parentTotalCount += count;
-			}
-		});
-
-		values.push(parentTotalCount);
-		values.push(...tempChildValues);
-	});
-
-	return { labels, values, parents };
-};
+import SunBurstChart, {
+	transformSunBurstData
+} from "../components/Charts/SunBurstChart";
 
 const Overview: NextPage = () => {
 	const queries = useQueries([
@@ -354,6 +324,7 @@ const Overview: NextPage = () => {
 											data={queryResults.modelsByDatasetAvailability.data}
 											dataEndPoint="dataset_available"
 											colors={datasetCountColors}
+											totalModelCount={queryResults.modelCount.data ?? 0}
 										/>
 									</Card>
 								</div>
